upgrade Html and versionUtils to TS
