+++ console/src/helpers/versionUtils.ts
@@ -10,25 +10,32 @@ const featureLaunchVersions = {
   [READ_ONLY_RUN_SQL_QUERIES]: 'v1.1.0',
 };
 
-export const checkValidServerVersion = version => {
+type IFeature = keyof typeof featureLaunchVersions;
+
+type IFeaturesCompatibility = {
+  [key in IFeature]?: boolean;
+};
+
+export const checkValidServerVersion = (version: string) => {
   return semver.valid(version) !== null;
 };
 
-export const getFeaturesCompatibility = serverVersion => {
-  const featuresCompatibility = {};
+export const getFeaturesCompatibility = (serverVersion: string) => {
+  const featuresCompatibility: IFeaturesCompatibility = {};
 
   const isValidServerVersion = checkValidServerVersion(serverVersion);
 
-  Object.keys(featureLaunchVersions).forEach(feature => {
+  Object.keys(featureLaunchVersions).forEach(_feature => {
+    const feature = _feature as IFeature;
     featuresCompatibility[feature] = isValidServerVersion
-      ? semver.satisfies(serverVersion, '>=' + featureLaunchVersions[feature])
+      ? semver.satisfies(serverVersion, `>= ${featureLaunchVersions[feature]}`)
       : true;
   });
 
   return featuresCompatibility;
 };
 
-export const versionGT = (version1, version2) => {
+export const versionGT = (version1: string, version2: string) => {
   try {
     return semver.gt(version1, version2);
   } catch (e) {
@@ -37,7 +44,7 @@ export const versionGT = (version1, version2) => {
   }
 };
 
-export const checkStableVersion = version => {
+export const checkStableVersion = (version: string) => {
   try {
     const preReleaseInfo = semver.prerelease(version);
 
@@ -48,7 +55,7 @@ export const checkStableVersion = version => {
   }
 };
 
-export const checkFeatureSupport = feature => {
+export const checkFeatureSupport = (feature: IFeature) => {
   return (
     globals.featuresCompatibility && globals.featuresCompatibility[feature]
   );
